Fix for Windows being nonstandard.
%zu for size_t is C99, but *no*, MSVC doesn't support it.

(Though technically it's incorrect to assume that size_type is size_t in
the first place, for some strange reason...)

+++ Source/Core/Core/Src/HW/USBInterface.cpp
@@ -320,7 +320,7 @@ void RegisterDeviceChangeClient(IUSBDeviceChangeClient* Client)
 void DeregisterDeviceChangeClient(IUSBDeviceChangeClient* Client)
 {
 	g_DeviceChangeClients.erase(Client);
-	DEBUG_LOG(USBINTERFACE, "DeregisterDeviceChangeClient(%p), now %zu", Client, g_DeviceChangeClients.size());
+	DEBUG_LOG(USBINTERFACE, "DeregisterDeviceChangeClient(%p), now %u", Client, (unsigned int) g_DeviceChangeClients.size());
 	if (g_ShouldScan && g_DeviceChangeClients.empty())
 	{
 		SetShouldScan(false);
